feat: close WMS Stage 42 (42.1-42.4), Stage 45 UI sweep, and Stage 46 money-paise migration
Stage 42.1 (lot/serial/UOM depth): LottableConstraint, UOM/UOMConversion masters,
Code 128 barcode rendering for the print-sheet fallback, plus a real bug fix
(ValidateMasterDataRules was unreachable from the generic doc-save path for
Batch/SerialNumber/LottableConstraint/UOMConversion) and a timezone bug fix in
GetFXGainLossRegister's default date window.

Stage 42.2 (task spine): WarehouseTask doctype + lifecycle, five retrofits
(PutawayToBin, ExecuteBinReplenishment, CrossDockPutaway, PostCycleCountAdjustment,
ScanPickItem) logging through one choke point, TaskDispatchStrategy, Zone/
PutawayStrategy/AllocationStrategy masters, bin capacity enforcement, exception
codes + follow-on actions, and the Warehouse Cockpit screen.

Stage 42.3 (inbound depth): DockDoor/Appointment scheduling, Yard Board,
Hold/HoldReleaseRequest maker-checker (closing a real RBAC bypass for HR/Admin),
receipt validation rule tolerances, catch-weight capture, planned cross-dock,
hazmat compatibility, compliance fields, RF Receiving.

Stage 42.4 (outbound depth): Wave planning/dispatch, Sortation, Cartonization v2,
Pack Station + packing validation templates, Deconsolidation, Loading + BOL,
Pre-Ship validation gate, Value-Added Services against the existing BOM engine.

Stage 45 (UI/UX design-bar sweep): fixed a render-race that could duplicate
panels (viewRenderToken + detached scratch container), corrected the frozen
F&A header regression, fixed 5 tab bars silently collapsing to 0 height under
overflow-x:auto, non-dismissible setup-prerequisite banner, Setup flyout
visual fixes, "Custom ERP" -> "Wholeops" rebrand, new Knowledge Center sidebar
entry.

Stage 46 (money precision): gl_postings.debit/credit widened from whole-rupee
INT to paise BIGINT (RupeesToPaise/PaiseToRupees/PaiseMap helpers), fixing GST
component floor-truncation on every posting and a cart-total truncation bug in
POS checkout and ProcessReturnAnywhere that dropped fractional-rupee unit
prices before multiplying by quantity. Every report/API boundary converts back
to rupees before leaving the function. Includes a data-mutating migration
(existing gl_postings rows are multiplied by 100 in place).

33 new additive migrations total. No new dependencies.

+++ engines/accounting_periods_test.go
@@ -36,7 +36,7 @@ func TestAccountingPeriodClosesBlockPostingsAndOpenPeriodsStillWork(t *testing.T
 	}
 
 	// While the period covering today is still Open, posting must succeed.
-	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "OPEN-CHECK", map[string]int{"1100": 100}, map[string]int{"4100": 100}, "", ""); err != nil {
+	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "OPEN-CHECK", map[string]int64{"1100": 100}, map[string]int64{"4100": 100}, "", ""); err != nil {
 		t.Fatalf("expected posting to succeed while period is Open: %v", err)
 	}
 
@@ -50,7 +50,7 @@ func TestAccountingPeriodClosesBlockPostingsAndOpenPeriodsStillWork(t *testing.T
 	}
 
 	// Now that the period covering today is Closed, posting must be rejected.
-	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "CLOSED-CHECK", map[string]int{"1100": 100}, map[string]int{"4100": 100}, "", ""); err == nil {
+	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "CLOSED-CHECK", map[string]int64{"1100": 100}, map[string]int64{"4100": 100}, "", ""); err == nil {
 		t.Fatalf("expected posting to be rejected while period is Closed")
 	}
 
@@ -59,13 +59,13 @@ func TestAccountingPeriodClosesBlockPostingsAndOpenPeriodsStillWork(t *testing.T
 	// date in this test - this exercises the $1::date branch specifically,
 	// not just the CURRENT_DATE default every other call in this codebase
 	// still uses.
-	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "BACKDATED-CHECK", map[string]int{"1100": 100}, map[string]int{"4100": 100}, today.Format("2006-01-02"), ""); err == nil {
+	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "BACKDATED-CHECK", map[string]int64{"1100": 100}, map[string]int64{"4100": 100}, today.Format("2006-01-02"), ""); err == nil {
 		t.Fatalf("expected posting with an explicit transactionDate inside the closed period to be rejected")
 	}
 	// And a transactionDate outside the closed period must still be allowed
 	// even though a closed period exists at all.
 	farFuture := today.AddDate(1, 0, 0).Format("2006-01-02")
-	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "FUTURE-DATE-CHECK", map[string]int{"1100": 100}, map[string]int{"4100": 100}, farFuture, ""); err != nil {
+	if err := PostDoubleEntry(tenantID, "TestPeriodDoc", "FUTURE-DATE-CHECK", map[string]int64{"1100": 100}, map[string]int64{"4100": 100}, farFuture, ""); err != nil {
 		t.Fatalf("expected posting with a transactionDate outside the closed period to succeed: %v", err)
 	}
